+++ src/contexts/Balances/types.ts
@@ -3,5 +3,8 @@ import BigNumber from "bignumber.js";
 export interface ContextValues {
   tokenBalances?: Array<BigNumber>;
   LPBalances?: Array<BigNumber>;
+  INKBalance?: BigNumber;
+  GRANSBalance?: BigNumber;
+  UBQBalance?: BigNumber;
   CurrentBlock?: string;
 }