+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@
 - added LP % for staked amount to the Farm page and shortened the displayed digits for easier reading
 - small UI visual style improvements
 - dev nice-to-have changes
+- added UBQ balance to My Walet modal
 - React 17 upgrade
 - added MUI components library
 - forked `use-wallet` repo to modify for react 17 and include UBQ network